src/requirements.txt
google-cloud-bigquery
google-cloud-storage
pandas
selenium
openpyxl
pyarrow
python-dateutil